Gestion-main/pyproject.toml
[project]
name = "gestion-main"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "amqp>=5.1.1",
    "appnope==0.1.3",
    "asgiref>=3.7.0",
    "asttokens==2.0.8",
    "attrs==22.1.0",
    "backcall==0.2.0",
    "bbox-visualizer==0.1.0",
    "billiard>=4.2.0",
    "celery>=5.3.0",
    "certifi==2022.9.24",
    "charset-normalizer==2.1.1",
    "click==8.1.3",
    "click-didyoumean==0.3.0",
    "click-plugins==1.1.1",
    "click-repl==0.2.0",
    "colorama==0.4.5",
    "contourpy>=1.2.0",
    "cycler==0.11.0",
    "debugpy==1.6.3",
    "decorator==5.1.1",
    "django>=4.2.0",
    "django-cors-headers==4.3.0",
    "djangorestframework==3.14.0",
    "djangorestframework-simplejwt==5.2.1",
    "entrypoints==0.4",
    "executing==1.1.0",
    "fonttools==4.37.4",
    "gunicorn==21.2.0",
    "idna==3.4",
    "ipykernel==6.16.0",
    "ipython==8.5.0",
    "ipywidgets==8.0.2",
    "jarowinkler>=1.3.0",
    "jedi==0.18.1",
    "joblib==1.2.0",
    "jupyter-bbox-widget==0.5.0",
    "jupyter-client==7.3.5",
    "jupyter-core==4.11.1",
    "jupyterlab-widgets==3.0.3",
    "kiwisolver>=1.4.5",
    "kombu>=5.3.0",
    "levenshtein>=0.23.0",
    "matplotlib>=3.8.0",
    "matplotlib-inline==0.1.6",
    "nest-asyncio==1.5.6",
    "numpy>=1.26.0",
    "opencv-python>=4.8.0",
    "packaging==21.3",
    "pandas>=2.1.0",
    "parso==0.8.3",
    "pexpect==4.8.0",
    "pickleshare==0.7.5",
    "pillow>=10.0.0",
    "prompt-toolkit==3.0.31",
    "psutil>=5.9.8",
    "psycopg2-binary",
    "psycopg[binary]",
    "ptyprocess==0.7.0",
    "pure-eval==0.2.2",
    "pygments==2.13.0",
    "pyjwt==2.5.0",
    "pylabels==1.2.1",
    "pyparsing==3.0.9",
    "python-barcode==0.14.0",
    "python-dateutil==2.8.2",
    "python-levenshtein>=0.23.0",
    "pytoml==0.1.21",
    "pytz>=2023.3",
    "pyyaml>=6.0.2",
    "pyzmq",
    "rapidfuzz>=3.0.0",
    "reportlab>=4.0.0",
    "requests==2.28.1",
    "scikit-learn>=1.3.0",
    "scipy>=1.11.0",
    "setuptools",
    "six==1.16.0",
    "sqlparse>=0.4.4",
    "stack-data==0.5.1",
    "threadpoolctl==3.1.0",
    "tornado==6.2",
    "traitlets==5.4.0",
    "tzdata>=2023.3",
    "urllib3==1.26.12",
    "vine==5.0.0",
    "wcwidth==0.2.5",
    "widgetsnbextension==4.0.3",
]
